pyproject.toml
[project]
name = "reposage"
version = "0.1.0"
description = "RAG-based developer tool for asking natural language questions about GitHub repositories"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Debdeep" }]

dependencies = [
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",

    # Validation & config
    "pydantic>=2.9",
    "pydantic-settings>=2.6",

    # GitHub
    "PyGithub>=2.5",
    "gitpython>=3.1",

    # Chunking — tree-sitter core + per-language grammars (official packages, arm64 wheels)
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
    "tree-sitter-go>=0.23",
    "tree-sitter-rust>=0.23",
    "tree-sitter-java>=0.23",
    "tree-sitter-cpp>=0.23",
    "tree-sitter-c>=0.23",
    "tree-sitter-ruby>=0.23",

    # Embeddings & reranking
    "sentence-transformers>=3.3",
    "torch>=2.5",
    "einops>=0.8",          # required by nomic-embed-text-v1.5

    # Tokenization (for chunk size budgeting)
    "tiktoken>=0.8",

    # Vector store
    "chromadb>=0.5.20",

    # LLM synthesis
    "anthropic>=0.40",

    # Async HTTP
    "httpx>=0.27",

    # Logging
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "ipython>=8.30",
]
eval = [
    "ragas>=0.2",
    "datasets>=3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["reposage"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
